Collect global context
Just collect in `ocs_page_context.js` file because the context is global and should be collected on every page.
