+++ corehq/apps/hqwebapp/static/hqwebapp/js/ocs_page_context.js
@@ -6,6 +6,9 @@
 *  Widget API: https://docs.openchatstudio.com/chat_widget/reference/#page-context
 */
 
+import $ from "jquery";
+import initialPageData from "hqwebapp/js/initial_page_data";
+
 const WIDGET_SELECTOR = 'open-chat-studio-widget';
 const BEFORE_SEND_EVENT = 'ocs:message:before-send';
 
@@ -15,9 +18,39 @@ function registerContextCollector(collectContext) {
     _contextCollectors.push(collectContext);
 }
 
+function _domainFromUrl() {
+    const match = window.location.pathname.match(/^\/a\/([^/]+)\//);
+    return match ? match[1] : null;
+}
+
+// Fetch once on load; role rarely changes and the endpoint is cached server-side.
+let _roleContext = {};
+
+function _fetchMyRole() {
+    let url;
+    try {
+        url = initialPageData.reverse('my_role');
+    } catch {
+        // my_role URL only exists on domain pages
+        return;
+    }
+    $.getJSON(url).done(function (data) {
+        _roleContext = data;
+    });
+}
+
+function collectGlobalContext() {
+    return Object.assign({
+        url: window.location.href,
+        page_title: document.title,
+        domain: _domainFromUrl(),
+    }, _roleContext);
+}
+
 function getClientPageContext() {
     return Object.assign(
         {},
+        collectGlobalContext(),
         ..._contextCollectors.map((collectContext) => collectContext() || {}),
     );
 }
@@ -27,6 +60,7 @@ document.addEventListener('DOMContentLoaded', function () {
     if (!widget) {
         return;
     }
+    _fetchMyRole();
     widget.addEventListener(BEFORE_SEND_EVENT, function () {
         widget.pageContext = getClientPageContext();
     });
